Squashed commit of the following:
commit 44cda8448db20e366d558e54525e17d9cbfe24f2
Author: Marcel Hruska <marcel.hruska@broadcom.com>
Date:   Tue Feb 4 20:55:48 2020 +0100

    Dynamic ports (#86)

    * small updates to package.json
    * added license
    * extension updates for version 0.9.0
    * dynamic ports
    * debug configurations now instance-specific
    * filename fix
    * implementation without get-port package
    * pull request suggestions

commit d783853d535c673eb0aa9a2db0524608494d9d38
Author: Michal Bali <michal.bali@broadcom.com>
Date:   Thu Jan 30 17:05:26 2020 +0100

    Language server comments (#81)

    * location counter in address fix
    * UB fix
    * language server tests for linux
    * comments
    * nullopt fix
    * alpine segfault fix
    * refactoring

+++ parser_library/src/debugging/debugger.cpp
@@ -144,7 +144,7 @@ const std::vector<stack_frame> & debugger::stack_frames()
 	stack_frames_.clear();
 	if (debug_ended_)
 		return stack_frames_;
-	for(size_t i = proc_stack_.size()-1; i != -1; --i)
+	for(size_t i = proc_stack_.size()-1; i != (size_t)-1; --i)
 	{
 		source source(proc_stack_[i].proc_location.file);
 		std::string name;
